move tables to a new file
